+++ CHANGELOG.md
@@ -28,6 +28,14 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 - this is an example of a deprecation note
 
 ---
+
+## [4.2] Support alpha mod in colors
+
+- Add `value/rgbComponents` transform to dictionary
+- Add `css/rgbComponents` format to dictionary
+- Add new distribution to dictionary, `lib/dictionary/css/rawRgbColors.css`
+
+
 ## [4.1] Removing some global button css
 
 - Removed global button line-height